requirements.txt
aiokafka==0.7.2
python-dotenv==0.19.2